Replace int64_t

+++ TTGDeallocTaskHelper/Classes/NSObject+TTGDeallocTaskHelper.m
@@ -37,13 +37,13 @@ const NSUInteger TTGDeallocTaskIllegalIdentifier = 0; // illegal identifier
 
 - (NSUInteger)addTask:(TTGDeallocTaskBlock)taskBlock {
     // Global increase identifier
-    static volatile int64_t globalIdentifier = 0;
+    static volatile NSUInteger globalIdentifier = 0;
     
     if (!taskBlock) {
         return TTGDeallocTaskIllegalIdentifier;
     }
     
-    int64_t newIdentifier = OSAtomicIncrement64(&globalIdentifier);
+    NSUInteger newIdentifier = OSAtomicIncrement64(&globalIdentifier);
     NSNumber *newIdentifierNumber = @(newIdentifier);
     
     if (newIdentifierNumber) {
